requirements.txt
face_recognition
dlib
opencv-python
numpy
paho-mqtt